refactor(cleanup): extract CleanupExecutor; move structural check onto plan
Completes the planner/executor split answering the PR #1574 architecture
review: common_nwbfile.py drops from 1893 to 1427 lines.

- Move the destructive act loop and its pre-delete guards
  (_candidate_still_matches, _containing_managed_root) into a schema-agnostic
  CleanupExecutor; _remove_untracked_files delegates to it. Guards move
  verbatim: whole-plan structural preflight, per-candidate tracking/registry
  refresh, act-time age recheck, alias/target/repoint/protected-root checks,
  per-leaf re-validation before unlink, fail-closed RuntimeError. The
  cleanup() coordinator is unchanged.
- Keep _current_tracked_state/_current_custom_tables/_other_managed_roots/
  _access_still_matches on AnalysisNwbfile (DataJoint-coupled and/or test
  seams) and inject them into the executor.
- Move the plan's structural-integrity check onto
  CleanupPlan.structural_problems() (kept act-time: the frozen dataclass's
  Set/Dict fields are mutable, so post-construction forging must be caught).
- Add logging.Logger / str|Path annotations; add direct fake-injection unit
  tests for CleanupPlanner and CleanupExecutor; fix a stale comment reference.

Full synthetic cleanup suite: 118 passed.
